Implemented game mechanics within the game loop
In the game loop we use a switch case to differentiate the game
state from the menu, level gameplay, and game over. Player movement is
taken from the keyboard. The player avatar and other game world objects
are checked and updated accordingly based on their defined controls
and/or collision handling.